Feature/#74 feat payment apply saga

+++ payment/src/main/java/com/onevoice/payment/presentation/dto/response/RefundPaymentResponse.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 
 public record RefundPaymentResponse(
     UUID refundId,
-    long refundAmount,
+    int refundAmount,
     RefundStatus refundStatus
 ) {
